SensorDataService: added hwpid and rssi

+++ src/DpaParser/EmbedFRC.h
@@ -34,7 +34,7 @@ namespace iqrf
         //params
         uint8_t m_frcCommand;
         std::vector<uint8_t> m_userData;
-        std::set<int> m_selectedNodes;
+        std::set<uint8_t> m_selectedNodes;
 
         //response
         int m_status;
@@ -46,7 +46,7 @@ namespace iqrf
           , m_status(0)
         {}
 
-        Send(uint8_t frcCommand, const std::set<int> & selectedNodes, const std::vector<uint8_t> & userData)
+        Send(uint8_t frcCommand, const std::set<uint8_t> &selectedNodes, const std::vector<uint8_t> &userData)
           :m_frcCommand(frcCommand)
           , m_userData(userData)
           , m_selectedNodes(selectedNodes)
@@ -58,7 +58,7 @@ namespace iqrf
           , m_status(0)
         {}
 
-        Send(const std::set<int> & selectedNodes)
+        Send(const std::set<uint8_t> &selectedNodes)
           :m_frcCommand(0xFF)
           , m_selectedNodes(selectedNodes)
           , m_status(0)
@@ -76,10 +76,10 @@ namespace iqrf
         const std::vector<uint8_t> & getFrcData() const { return m_frcData; }
 
         // get selected nodes
-        const std::set<int> & getSelectedNodes() const { return m_selectedNodes; }
+        const std::set<uint8_t> & getSelectedNodes() const { return m_selectedNodes; }
 
         // makes selected cmd
-        void setSelectedNodes(const std::set<int> & val)
+        void setSelectedNodes(const std::set<uint8_t> &val)
         {
           m_selectedNodes = val;
         }
